whisperAI/requirements.txt
openai-whisper==20231117
torch --index-url https://download.pytorch.org/whl/cu124
torchvision --index-url https://download.pytorch.org/whl/cu124
torchaudio --index-url https://download.pytorch.org/whl/cu124
numpy
grpcio
librosa
scipy
protobuf==5.27.2